+++ src/structures/Command.ts
@@ -18,4 +18,18 @@ export class Command {
 	run(interaction: any, additionalParams?: any) {
 		throw new MethodNotOverridenError(`run() method on ${this.name} command is not present.`);
 	}
+
+	toJSON() {
+		return {
+			name: this.name,
+			description: this.description,
+			options: this.options,
+			default_permissions: this.defaultPermissions,
+			disabled: this.disabled,
+			category: this.category,
+			guild_id: this.guildId,
+			user_ids: this.userIds,
+			guild_ids: this.guildIds,
+		};
+	}
 }